refactor: custom hook to preload an asset
