+++ packages/shared/src/hooks/useOnboardingModal.ts
@@ -7,6 +7,7 @@ import { AnalyticsEvent } from '../lib/analytics';
 import { OnboardingVersion } from '../lib/featureValues';
 import { cloudinary } from '../lib/image';
 import { LoggedUser } from '../lib/user';
+import { AssetType, useAssetPreload } from './useAssetPreload';
 import { useMyFeed } from './useMyFeed';
 import usePersistentContext from './usePersistentContext';
 
@@ -38,6 +39,7 @@ export const useOnboardingModal = ({
   onboardingVersion,
   onFeedPageChanged,
 }: UseOnboardingModalProps): UseOnboardingModal => {
+  useAssetPreload(AssetType.image, cloudinary.feedFilters.yourFeed);
   const { trackEvent } = useContext(AnalyticsContext);
   const { registerLocalFilters } = useMyFeed();
   const [shouldUpdateFilters, setShouldUpdateFilters] = useState(false);
@@ -87,18 +89,6 @@ export const useOnboardingModal = ({
     setOnboardingMode(OnboardingMode.Auto);
     modalStateCommand[onboardingVersion]?.(true);
   }, [hasOnboardingLoaded, user]);
-
-  useEffect(() => {
-    const link = document.createElement('link');
-    link.setAttribute('rel', 'preload');
-    link.setAttribute('href', cloudinary.feedFilters.yourFeed);
-    link.setAttribute('as', 'image');
-    document.head.append(link);
-
-    return () => {
-      link.remove();
-    };
-  }, []);
 
   return useMemo(
     () => ({
